Refactor: Separate constants into domain-specific files
Moved constants from a single file into multiple files based on their domain (events, commands, storage, etc.) to improve organization and maintainability.

- Created new files under `src/common/constants/` for each category.
- Created a barrel file `src/common/constants/index.ts` to re-export all constants.
- Updated all import statements throughout the codebase to use the new structure.
- Removed the old `src/common/constants.ts` file.
- Ensured the application builds successfully after changes.
