+++ package.json
@@ -521,7 +521,7 @@
     "@vscode/vsce": "^3.1.1",
     "copy-webpack-plugin": "^12.0.2",
     "css-loader": "^6.8.1",
-    "esbuild": "^0.21.4",
+    "esbuild": "^0.21.5",
     "esbuild-plugin-copy": "^2.1.1",
     "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-simple-import-sort": "^12.1.1",
