+++ src/common/constants/commands.ts
@@ -1,0 +1,28 @@
+export const TWINNY_COMMAND_NAME = {
+  addTests: "twinny.addTests",
+  addTypes: "twinny.addTypes",
+  conversationHistory: "twinny.conversationHistory",
+  disable: "twinny.disable",
+  embeddings: "twinny.embeddings",
+  enable: "twinny.enable",
+  review: "twinny.review",
+  explain: "twinny.explain",
+  focusSidebar: "twinny.sidebar.focus",
+  generateDocs: "twinny.generateDocs",
+  getGitCommitMessage: "twinny.getGitCommitMessage",
+  hideBackButton: "twinny.hideBackButton",
+  manageProviders: "twinny.manageProviders",
+  manageTemplates: "twinny.manageTemplates",
+  newConversation: "twinny.newConversation",
+  openPanelChat: "twinny.openPanelChat",
+  openChat: "twinny.openChat",
+  refactor: "twinny.refactor",
+  settings: "twinny.settings",
+  stopGeneration: "twinny.stopGeneration",
+  templateCompletion: "twinny.templateCompletion",
+  templates: "twinny.templates",
+  twinnySymmetryTab: "twinny.symmetry",
+  addFileToContext: "twinny.addFileToContext",
+  getContextFiles: "twinny.getContextFiles",
+  addSelectionToContext: "twinny.addSelectionToContext"
+}
