+++ src/common/constants/context.ts
@@ -1,0 +1,21 @@
+export const EXTENSION_CONTEXT_NAME = {
+  twinnyConversationHistory: "twinnyConversationHistory",
+  twinnyEnableRag: "twinnyEnableRag",
+  twinnyGeneratingText: "twinnyGeneratingText",
+  twinnyManageProviders: "twinnyManageProviders",
+  twinnyManageTemplates: "twinnyManageTemplates",
+  twinnyMaxChunkSize: "twinnyMaxChunkSize",
+  twinnyMinChunkSize: "twinnyMinChunkSize",
+  twinnyOverlapSize: "twinnyOverlapSize",
+  twinnyRelevantCodeSnippets: "twinnyRelevantCodeSnippets",
+  twinnyRelevantFilePaths: "twinnyRelevantFilePaths",
+  twinnyRerankThreshold: "twinnyRerankThreshold",
+  twinnyReviewTab: "twinnyReviewTab",
+  twinnySymmetryTab: "twinnySymmetryTab",
+  twinnyEmbeddingsTab: "twinnyEmbeddingsTab"
+}
+
+export const EXTENSION_SESSION_NAME = {
+  twinnySymmetryConnection: "twinnySymmetryConnection",
+  twinnySymmetryConnectionProvider: "twinnySymmetryConnectionProvider"
+}
