+++ src/common/constants/ui.ts
@@ -1,0 +1,9 @@
+export const WEBUI_TABS = {
+  chat: "chat",
+  history: "history",
+  providers: "providers",
+  review: "review",
+  settings: "templates",
+  symmetry: "symmetry",
+  embeddings: "embeddings"
+}
